backend/requirements.txt
asgiref==3.5.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
certifi==2021.10.8; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cffi==1.15.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3"
coreapi==2.3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
coreschema==0.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cryptography==37.0.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
defusedxml==0.7.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
django-colorfield==0.6.3
django-debug-toolbar==3.4.0; python_version >= "3.7"
django-templated-mail==1.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
django==3.2.13; python_version >= "3.6"
djangorestframework-simplejwt==4.8.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
djoser==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
idna==3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.5"
importlib-metadata==1.7.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version < "3.8"
itypes==1.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
jinja2==3.1.2; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
markupsafe==2.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
oauthlib==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pillow==9.1.0; python_version >= "3.7"
psycopg2==2.9.3; python_version >= "3.6"
pycparser==2.21; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pyjwt==2.3.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
python-dotenv==0.20.0; python_version >= "3.5"
python3-openid==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pytz==2022.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
requests-oauthlib==1.3.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
requests==2.27.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-app-django==4.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-core==4.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.7"
typing-extensions==4.2.0; python_version < "3.8" and python_version >= "3.7" and python_full_version >= "3.6.1" and python_full_version < "4.0.0"
uritemplate==4.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
urllib3==1.26.9; python_full_version >= "3.6.1" and python_version < "4" and python_full_version < "4.0.0"
zipp==3.8.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version < "3.8" and python_version >= "3.7"
